requirements.txt
pinecone-client>=2.2.4
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0




